modif CI failed

+++ paddle/fluid/operators/sequence_ops/sequence_mask_op_npu.cc
@@ -58,9 +58,9 @@ class SequenceMaskNPUKernel : public framework::OpKernel<T> {
 
     Tensor cast_x;
     cast_x.mutable_data<int32_t>(x->dims(), ctx.GetPlace());
-    const auto& cast1_runner =
-        NpuOpRunner("Cast", {*x}, {cast_x},
-                    {{"dst_type", ConvertToNpuDtype(cast_x.type())}});
+    const auto& cast1_runner = NpuOpRunner(
+        "Cast", {*x}, {cast_x},
+        {{"dst_type", static_cast<int>(ConvertToNpuDtype(cast_x.type()))}});
     cast1_runner.Run(dev_ctx.stream());
 
     Tensor tmp;
@@ -120,7 +120,8 @@ class SequenceMaskNPUKernel : public framework::OpKernel<T> {
     }
 
     const auto& cast2_runner = NpuOpRunner(
-        "Cast", {y_tmp}, {*y}, {{"dst_type", ConvertToNpuDtype(out_dtype)}});
+        "Cast", {y_tmp}, {*y},
+        {{"dst_type", static_cast<int>(ConvertToNpuDtype(out_dtype))}});
     cast2_runner.Run(dev_ctx.stream());
   }
 };
